+++ src/cmenu.cpp
@@ -38,7 +38,7 @@ int eContextMenu::event(eEvent& ev) {
       if (ev.coord.x>x && ev.coord.x<x+w &&
           ev.coord.y>y && ev.coord.y<y+h) {
         selected=(ev.coord.y-y-2)/18;
-        if (selected>=items.size()) {
+        if (selected>=(int)items.size()) {
           selected=-1;
         }
       } else {
@@ -66,11 +66,11 @@ int eContextMenu::draw() {
     engine->frect(x+1,(1.5+y+(selected*18)),x+w-1.5,(1.5+y+(1+selected)*18));
   }
   engine->drawColor({1,1,1,1});
-  for (int i=0; i<items.size(); i++) {
+  for (size_t i=0; i<items.size(); i++) {
     if (items[i].dtext==NULL) {
       items[i].dtext=new sf::Text(items[i].text,engine->defFont->inst,12*engine->scale);
     }
-    items[i].dtext->setPosition((4+x)*engine->scale,(3+y+i*18)*engine->scale);
+    items[i].dtext->setPosition((4+x)*engine->scale,(3+y+(int)i*18)*engine->scale);
     engine->win->draw(*items[i].dtext);
   }
   return 1;
